clone.rs chdir.rs clone_repo.rs execute.rs execution.rs

+++ src/commands/clone.rs
@@ -6,7 +6,7 @@ pub fn clone(org: &str) -> Vec<Step> {
     let mut result = vec![];
     let repos = github::get_repos(org);
     for (i, repo) in repos.into_iter().enumerate() {
-        result.push(operations::clone_repo(i + 1, repo.clone_url));
+        result.push(operations::clone_repo((i as u32) + 1, repo.clone_url));
     }
     result
 }
