+++ src/operations/chdir.rs
@@ -1,5 +1,5 @@
 use crate::runtime::Step;
 
-pub fn chdir(id: usize, dir: String) -> Step {
+pub fn chdir(id: u32, dir: String) -> Step {
     Step::Chdir { id, dir }
 }
