+++ src/operations/clone_repo.rs
@@ -1,7 +1,7 @@
 use crate::runtime::Step;
 
 /// provides the Step for cloning the given repo
-pub fn clone_repo(id: usize, url: String) -> Step {
+pub fn clone_repo(id: u32, url: String) -> Step {
     Step::Run {
         id,
         command: "git".into(),
